# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Room Catalog Service API",
        "description": "Service Katalog & Kamar - Smart Hospitality IAE Kelompok 11",
        "contact": {
            "email": "admin@hotel.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Room Catalog Service"
        }
    ],
    "paths": {
        "/api/v1/rooms": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Mengambil daftar seluruh kamar",
                "operationId": "8621bf9d381104070fb98293d57bc3c2",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "available",
                                "occupied",
                                "maintenance"
                            ]
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "standard",
                                "deluxe",
                                "suite",
                                "presidential"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar kamar berhasil diambil"
                    }
                },
                "security": [
                    {
                        "X-IAE-KEY": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Menambah data kamar baru (triggers SSO + SOAP audit + RabbitMQ)",
                "operationId": "49058005cbeb1d687de909882996a7f7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "room_number",
                                    "type",
                                    "floor",
                                    "capacity",
                                    "price_per_night"
                                ],
                                "properties": {
                                    "room_number": {
                                        "type": "string",
                                        "example": "101"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "standard",
                                            "deluxe",
                                            "suite",
                                            "presidential"
                                        ]
                                    },
                                    "floor": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "price_per_night": {
                                        "type": "number",
                                        "example": 500000
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "facilities": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kamar berhasil ditambahkan"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "X-IAE-KEY": []
                    }
                ]
            }
        },
        "/api/v1/rooms/{id}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Mengambil data spesifik kamar berdasarkan ID",
                "operationId": "cbfb744a4a46d46906c8f6aa65e0e452",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data kamar berhasil diambil"
                    },
                    "404": {
                        "description": "Kamar tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "X-IAE-KEY": []
                    }
                ]
            }
        },
        "/api/v1/rooms/{id}/status": {
            "put": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Mengubah status kamar",
                "operationId": "730022709fab2f4645c8499ed6c3226d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "available",
                                            "occupied",
                                            "maintenance"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status kamar berhasil diperbarui"
                    },
                    "404": {
                        "description": "Kamar tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "X-IAE-KEY": []
                    }
                ]
            }
        },
        "/api/v1/rooms/{id}/assign": {
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Menetapkan kamar ke tamu (triggers SSO + SOAP audit + RabbitMQ)",
                "operationId": "af9d1fe699defbdbcac4c0c811b8a688",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "guest_name",
                                    "reservation_id"
                                ],
                                "properties": {
                                    "guest_name": {
                                        "type": "string",
                                        "example": "Budi Santoso"
                                    },
                                    "reservation_id": {
                                        "type": "string",
                                        "example": "RES-2024-001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kamar berhasil di-assign"
                    },
                    "409": {
                        "description": "Kamar sudah terisi"
                    },
                    "404": {
                        "description": "Kamar tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "X-IAE-KEY": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Room": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "room_number": {
                        "type": "string",
                        "example": "101"
                    },
                    "type": {
                        "type": "string",
                        "example": "deluxe"
                    },
                    "floor": {
                        "type": "integer",
                        "example": 1
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 2
                    },
                    "price_per_night": {
                        "type": "number",
                        "example": 500000
                    },
                    "status": {
                        "type": "string",
                        "example": "available"
                    },
                    "description": {
                        "type": "string",
                        "example": "Kamar deluxe dengan view kolam renang"
                    },
                    "facilities": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2024-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "X-IAE-KEY": {
                "type": "apiKey",
                "name": "X-IAE-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Rooms",
            "description": "Rooms"
        }
    ]
}